// src/types/carrier/transport.ts
export type ITrailer = {
  created_at: string;
  cut_off_time: string; // ISO 8601 Date String
  trailer_license_plt_num_id: string;
  cns_line_item_by_date_id: number;
  price: string;
  matching_id: string;
  matching_status: string;
  matching_type: string;
  order_id: string;
  order_status: string;
  vehicle_avb_resource_id: string;
  total_count: string;
  trans_type: number;
  temperature_range: string[];
  trip_name: string;
  vehicle_name: string;
  vehicle_diagram_item_id: string | null;
  shipper_operator_id: string;
  carrier_operator_id: string;
  carrier2_operator_id: string;
  display_order: number | null;
  cut_off_fee: number;
  cut_off_info: ICutOffInfo[];
  departure_time_min: string;
  departure_time_max: string;
  is_emergency: boolean;
};

export type ICutOffInfo = {
  id: string;
  operatorId: number;
  cutOffTime: number;
  cutOffFee: number;
  created_user: string;
  vehicleAvbResourceId: number;
};
// ...
export interface IMatchingItem {
  id: number;
  matching_id: number;
  weight: number | null;
  status: number;
  operator_id: string;
  operator_code: string | null;
  req_cns_line_item_id: number | null;
  trans_plan_id: number | null;
  transport_code: string;
  transport_name: string;
  collection_date: string;
  collection_time_from: string;
  collection_time_to: string;
  departure_from: number;
  arrival_to: number;
  trailer_number: number;
  trailer_number_rest: number;
  price_per_unit: number;
  cut_off_fee: number;
  vehicle_condition: string[];
  temperature_range: [];
  outer_package_code: number;
  total_length: number | null;
  total_width: number | null;
  total_height: number | null;
  special_instructions: string | null;
  trans_type: number;
  trans_order_id: number | null;
  cns_line_item_id: number | null;
  shipper_operator_id: string;
  shipper_operator_name: string;
  vehicle_avb_resource_item_id: number;
  parent_order_propose_snapshot: any;
  operator_name: string;
  trailer_license_plt_num_id: string;
  created_at: string;
  carrier_operator_id: string;
  carrier_operator_name: string;
  carrier2_operator_id: string;
  carrier2_operator_name: string;
}